Fix a bug in Remove-ItemSilently by having retry logic when removing directories recursively

+++ .build/scripts/Pask.ps1
@@ -233,10 +233,18 @@ function script:Remove-ItemSilently {
 
         $ItemObject = Get-Item -Path $Item
 
-        if ($ItemObject -is [System.IO.FileInfo]) {
+        if ($ItemObject -is [System.IO.FileInfo] -and (Test-Path $ItemObject.FullName)) {
             Remove-Item -Path $ItemObject.FullName -Force | Out-Null
-        } elseif ($ItemObject -is [System.IO.DirectoryInfo]) {
-            [System.IO.Directory]::Delete(("{0}" -f $ItemObject.FullName), $true)
+        } elseif ($ItemObject -is [System.IO.DirectoryInfo] -and (Test-Path $ItemObject.FullName)) {
+            $Count = 0
+            $Retry = 10
+            do {
+                $Count++
+                CMD /C ("RD /S /Q ""{0}""" -f $ItemObject.FullName) 2>&1 | Out-Null
+            } while ((Test-Path $ItemObject.FullName) -and $Count -le $Retry)
+            if ($Count -gt $Retry) {
+                CMD /C ("RD /S /Q ""{0}""" -f $ItemObject.FullName)
+            }
         } else {
             $ItemObject.FullName | Remove-ItemSilently
         }
